+++ Rinkudesu.Identity.Service/Program.cs
@@ -12,12 +12,12 @@ using Microsoft.AspNetCore.Mvc.Versioning;
 using Microsoft.EntityFrameworkCore;
 using Microsoft.EntityFrameworkCore.Diagnostics;
 using Microsoft.OpenApi.Models;
-using Rinkudesu.Gateways.Webui.Models;
 using Rinkudesu.Identity.Service.Data;
 using Rinkudesu.Identity.Service.Middleware;
 using Rinkudesu.Identity.Service.Models;
 using Rinkudesu.Identity.Service.Repositories;
 using Rinkudesu.Identity.Service.Services;
+using Rinkudesu.Identity.Service.Settings;
 using Rinkudesu.Identity.Service.Utilities;
 using Rinkudesu.Kafka.Dotnet;
 using Rinkudesu.Kafka.Dotnet.Base;
