+++ Rinkudesu.Identity.Service/Settings/RedisSettings.cs
@@ -1,9 +1,8 @@
 ﻿using System.Diagnostics.CodeAnalysis;
 using Microsoft.Extensions.Caching.StackExchangeRedis;
-using Microsoft.Extensions.DependencyModel;
 using Rinkudesu.Identity.Service.Utilities;
 
-namespace Rinkudesu.Gateways.Webui.Models;
+namespace Rinkudesu.Identity.Service.Settings;
 
 /// <summary>
 /// Defines settings used for connection to Redis cache.
