+++ Rinkudesu.Identity.sln
@@ -6,6 +6,8 @@ Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Rinkudesu.Identity.Service.
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Rinkudesu.Identity.Service.MessageQueues", "Rinkudesu.Identity.Service.MessageQueues\Rinkudesu.Identity.Service.MessageQueues.csproj", "{82D5F806-C283-4557-B3EA-5CEE92CD2EE5}"
 EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Rinkudesu.Identity.Service.Models", "Rinkudesu.Identity.Service.Models\Rinkudesu.Identity.Service.Models.csproj", "{0D8CD85C-8AA9-46DB-96DA-E2DB1EE9B297}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -24,5 +26,9 @@ Global
 		{82D5F806-C283-4557-B3EA-5CEE92CD2EE5}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{82D5F806-C283-4557-B3EA-5CEE92CD2EE5}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{82D5F806-C283-4557-B3EA-5CEE92CD2EE5}.Release|Any CPU.Build.0 = Release|Any CPU
+		{0D8CD85C-8AA9-46DB-96DA-E2DB1EE9B297}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{0D8CD85C-8AA9-46DB-96DA-E2DB1EE9B297}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{0D8CD85C-8AA9-46DB-96DA-E2DB1EE9B297}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{0D8CD85C-8AA9-46DB-96DA-E2DB1EE9B297}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 EndGlobal
